+++ lib/constants.py
@@ -1,2 +1,6 @@
 PLUGIN_NAME = "NAD Kaarten"
 PLUGIN_ID = "nad_maps"
+# ADMIN_USERNAMES = ['svanderhoeven']
+# ADMIN_USERNAMES = ['Stijn.Overmeen']
+ADMIN_USERNAMES = ['']
+SERVICE_ENDPOINT = "https://api.pdok.nl/bzk/locatieserver/search/v3_1"
