+++ lib/thema.py
@@ -4,32 +4,91 @@
 from qgis.PyQt.QtGui import QStandardItem, QStandardItemModel
 import os
 import json
-from qgis.PyQt.QtCore import Qt, QRegularExpression
-from qgis.core import QgsProject
-from qgis.PyQt import QtCore
-from .load_layers import load_thema_layer
+from qgis.PyQt.QtCore import Qt, QRegularExpression, QSortFilterProxyModel, QItemSelectionModel
+from qgis.core import QgsProject, QgsRasterLayer, QgsVectorLayer
+from qgis.PyQt.QtWidgets import QAbstractItemView
+# from .load_layers import load_thema_layer
+from .style import StyleManager
+
 
 class ThemaManager:
     """
     Class to manage the thema sets (a list of one or more map layers)
     """
-    def __init__(self, dlg, plugin_dir, working_dir, creator):
+    def __init__(self, dlg, popup, plugin_dir, working_dir, creator, log):
 
         assert dlg is not None, "ThemaManager: dlg is None"
+        assert popup is not None, "ThemaManager: popup is None"
         assert plugin_dir is not None, "ThemaManager: plugin_dir is None"
         assert working_dir is not None, "ThemaManager: working_dir is None"
         assert creator is not None, "ThemaManager: creator is None"
+        assert log is not None, "ThemaManager: log is None"
 
         self.plugin_thema_path = os.path.join(plugin_dir, "resources", "themas", "thema.json")
+        self.plugin_styling_path = os.path.join(plugin_dir, "resources", "styling", "styling.json")
+        self.plugin_styling_files_path = os.path.join( plugin_dir, "resources", "styling", "qml_files")
         assert os.path.exists(self.plugin_thema_path), f"ThemaManager: plugin_thema_path does not exist: {self.plugin_thema_path}"
 
         self.set_working_directory(working_dir)
         
         self.dlg = dlg
+        self.creator = creator
+        self.log = log
+        self.style_manager = StyleManager(dlg=self.dlg, popup=popup, plugin_dir=plugin_dir, working_dir=working_dir, creator=creator, log=self.log)
+
         self.current_thema = None
         self.current_layer = None
         self.themaModel = QStandardItemModel()
         self.dlg.themaView.setModel(self.themaModel)
+
+        self.themaModel = QStandardItemModel()
+        self.favoriteFilterThema = QSortFilterProxyModel()
+        self.favoriteFilterThema.setSourceModel(self.themaModel)
+        self.favoriteFilterThema.setFilterKeyColumn(1)
+        self.favoriteFilterThema.setFilterRole(Qt.CheckStateRole) # https://doc.qt.io/qtforpython-6/PySide6/QtCore/QSortFilterProxyModel.html#PySide6.QtCore.QSortFilterProxyModel.setFilterRole
+
+        self.userFilterThema = QSortFilterProxyModel()
+        self.userFilterThema.setSourceModel(self.favoriteFilterThema)
+        self.userFilterThema.setFilterKeyColumn(2) # change this when you want to order by something else (like order in layer panel)
+
+        self.dlg.themaView.setModel(self.userFilterThema)
+        self.dlg.themaView.setEditTriggers(QAbstractItemView.EditTrigger.NoEditTriggers)
+        self.themaViewSelectionModel = QItemSelectionModel(self.dlg.themaView.model())
+        
+        self.themaMapModel = QStandardItemModel()
+        
+        self.proxyModelThemaMaps = QSortFilterProxyModel()
+        self.proxyModelThemaMaps.setSourceModel(self.themaMapModel)
+
+        self.dlg.themaMapListView.setModel(self.proxyModelThemaMaps)
+        self.dlg.themaMapListView.setEditTriggers(QAbstractItemView.EditTrigger.NoEditTriggers)
+
+
+        self.dlg.themaView.clicked.connect(
+            lambda cell: self.update_favorites(cell)
+        )
+        # Update the display with a list of map layers within the selected thema
+        self.dlg.themaView.selectionModel().selectionChanged.connect(
+            self.show_thema_layers
+        )
+        # TODO: what does this do?
+        self.dlg.themaView.selectionModel().select(
+            self.themaModel.index(0,0), QItemSelectionModel.Select | QItemSelectionModel.Rows
+        )
+        
+        self.dlg.themaView.doubleClicked.connect(
+            lambda: self.load_thema_layers()
+        )  # Using lambda here to prevent sending signal parameters to the loadService() function
+
+        self.service_type_mapping = {
+            "wms": "WMS",
+            "wmts": "WMTS",
+            "wfs": "WFS",
+            "wcs": "WCS",
+            "api features": "OGC API - Features",
+            "api tiles": "OGC API - Tiles",
+        }
+
         self.layer_type_mapping = {
             0: "Vector",
             1: "Raster",
@@ -43,6 +102,7 @@ class ThemaManager:
             9: "PostGIS",
             10: "Database"
         }
+        # self.log("Finished init ThemaManager")
 
     def set_working_directory(self, path):
         """Set the working directory for the plugin"""
@@ -55,14 +115,10 @@ class ThemaManager:
         os.makedirs(path, exist_ok=True)
         os.makedirs(os.path.join(path, "themas"), exist_ok=True)
 
-        self.user_thema_path = os.path.join(
-            path,
-            "themas", "user_themas.json"
-        )
-        self.user_thema_favorite_path = os.path.join(
-            path,
-            "themas", "favorites.json"
-        )
+        self.user_thema_path = os.path.join( path, "themas", "user_themas.json")
+        self.user_thema_favorite_path = os.path.join( path, "themas", "favorites.json")
+        self.user_styling_path = os.path.join(path, "styling", "styling.json")
+        self.user_styling_files_path = os.path.join(path, "styling", "qml_files")
 
     def delete_thema(self):
         """Delete an existing thema (only user defined themas should be deleted)"""
@@ -91,7 +147,7 @@ class ThemaManager:
 
         self.update_thema_list()
 
-    def save_thema(self, all: bool):
+    def save_thema(self, all: bool, selected_active_layers):
         """
         Save a collection of layers in order to later quickly load them
         """
@@ -110,7 +166,7 @@ class ThemaManager:
         # https://doc.qt.io/qt-6/qabstractitemmodel.html#details # documentation p1
         # https://doc.qt.io/qt-6/qt.html#ItemDataRole-enum # documentation p2
         if all == False:
-            selected_layers = self.selected_active_layers
+            selected_layers = selected_active_layers
         else:
             # selected_layers = QgsProject.instance().mapLayers().values()
             root = QgsProject.instance().layerTreeRoot()
@@ -120,6 +176,7 @@ class ThemaManager:
         # self.log(f"List of active selected layers is {selected_layers}, with length is {len(selected_layers)}")
         for i, layer in enumerate(selected_layers):
             layer_type = self.layer_type_mapping[layer.type()]
+            # self.log(f"layer.type is {layer.type()} and layer_type is {layer_type}")
             styling = layer.customProperty( 'layerStyle', '' )
             # self.log(f"Info of layer {layer.name()}, styling is {layer.customProperty( "layerStyle", "" )}, provider_type: {layer.providerType()}, layer_type: {layer_type}, source: {layer.source()}")
             string = f"{string}\"name\": \"{layer.name()}\"," # layer name
@@ -191,14 +248,17 @@ class ThemaManager:
                 favorites = json.load(f)
         except:
             favorites = None
-        
+    
         themas_exist = False
         for thema in themas:
             itemThema = QStandardItem(str(thema["thema_name"]))
             itemFavorite = QStandardItem()
             itemFavorite.setCheckable(True)
-            if favorites and favorites[thema["thema_name"]] == "favorite":
-                itemFavorite.setCheckState(2)
+            try:
+                if favorites and favorites[thema["thema_name"]] == "favorite":
+                    itemFavorite.setCheckState(2)
+            except:
+                itemFavorite.setCheckState(0)
             itemSource = QStandardItem(str(thema["creator"]))
             itemFilter = QStandardItem(f'{thema["thema_name"]} {thema["layers"]}')
             # https://doc.qt.io/qt-6/qstandarditem.html#setData
@@ -211,10 +271,12 @@ class ThemaManager:
         # if no thema is leftover after selection, then present an empty row
         if not themas_exist:
             itemThema = QStandardItem(str(""))
+            itemFavorite = QStandardItem()
+            itemFavorite.setCheckable(True)
             itemSource = QStandardItem(str(""))
             itemFilter = QStandardItem(str(""))
             self.themaModel.appendRow(
-                [itemThema, itemSource, itemFilter]
+                [itemThema, itemFavorite, itemSource, itemFilter]
             )
 
         self.themaModel.setHeaderData(2, Qt.Orientation.Horizontal, "Bron")
@@ -229,7 +291,7 @@ class ThemaManager:
             0, 300
         )  # set name to 300px (there are some huge layernames)
         self.dlg.themaView.horizontalHeader().setStretchLastSection(True)
-        self.dlg.themaView.sortByColumn(0, QtCore.Qt.AscendingOrder)
+        self.dlg.themaView.sortByColumn(0, Qt.AscendingOrder)
 
     def update_favorites(self, cell):
         if cell.column() == 1:
@@ -241,7 +303,7 @@ class ThemaManager:
                 thema_name = thema.data()
                 favorite = model.index(r, 1)
                 value = favorite.data(Qt.ItemDataRole.CheckStateRole) # https://doc.qt.io/qt-6/qt.html#CheckState-enum
-                self.log(f"Update favorites: CheckStateRole is {value} and DisplayRole is {thema.data(Qt.ItemDataRole.DisplayRole)}")
+                # self.log(f"Update favorites: CheckStateRole is {value} and DisplayRole is {thema.data(Qt.ItemDataRole.DisplayRole)}")
                 if value == 2:
                     checkstate = "favorite"
                 else:
@@ -259,8 +321,6 @@ class ThemaManager:
                     json.dump(data, file, indent='\t')
             except Exception as e:
                 self.log(f"Tried to write favorite themas to json file, but received this error: {e}")
-
-
 
 
 #########################################################################################
@@ -296,14 +356,31 @@ class ThemaManager:
             style = layer["styling"]
             # title, layer_type, provider_type, uri
             self.current_layer = layer
-            result = load_thema_layer(name, uri, layer_type, provider_type)
+            result = self.load_thema_layer(name, uri, layer_type, provider_type)
             QgsProject.instance().addMapLayer(result, False) # If True (by default), the layer will be added to the legend and to the main canvas
             group.addLayer(result) # Add the layer to the group
             # check if styling is saved in .resources.styling styling.json
             # if so, then apply that style, else apply no style
 
             if not style == "":
-                style_code = self.style_code(style, uri)
+
+                layer_style_list = []
+                # Load plugin styles
+                try:
+                    with open(self.plugin_styling_path, "r", encoding="utf-8") as f:
+                        layer_style_list.extend(json.load(f))
+                except:
+                    pass
+
+                # Load user styles
+                if os.path.exists(self.user_styling_path):
+                    try:
+                        with open(self.user_styling_path, "r", encoding="utf-8") as f:
+                            layer_style_list.extend(json.load(f))
+                    except:
+                        pass
+
+                style_code = self.style_manager.style_code(style, uri)
                 path = f"{self.plugin_styling_files_path}/{style_code}.qml"
                 result.loadNamedStyle(path)
                 result.triggerRepaint()
@@ -311,6 +388,16 @@ class ThemaManager:
 
         # self.iface.layerTreeView().collapseAllNodes()
         # self.update_active_layers_list()
+
+    # This functions loads a layer where the uri is already known and saved to a thema
+    def load_thema_layer(self, title, uri, layer_type, provider_type):
+        "Create and load a layer from the thema layer-list"
+        if layer_type == "Vector":
+            result = QgsVectorLayer(uri, title, provider_type)
+        elif layer_type == "Raster":
+            result = QgsRasterLayer(uri, title, provider_type)
+        # return QgsProject.instance().addMapLayer(result, True)
+        return result
 
     def update_thema_layers(self):
         """Update the list of layers contained with this thema"""
@@ -342,7 +429,7 @@ class ThemaManager:
                 )
         
         self.themaMapModel.setHeaderData(3, Qt.Orientation.Horizontal, "Bron")
-        self.themaMapModel.setHeaderData(2, Qt.Orientation.Horizontal, "Style")
+        self.themaMapModel.setHeaderData(2, Qt.Orientation.Horizontal, "Opmaak")
         self.themaMapModel.setHeaderData(1, Qt.Orientation.Horizontal, "Type")
         self.themaMapModel.setHeaderData(0, Qt.Orientation.Horizontal, "Laagnaam")
         self.themaMapModel.horizontalHeaderItem(3).setTextAlignment(Qt.AlignmentFlag.AlignLeft)
